+++ sdk/core/azure-core-http-netty/src/main/java/com/azure/core/http/netty/NettyAsyncHttpClient.java
@@ -27,7 +27,9 @@ import com.azure.core.util.logging.ClientLogger;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 import io.netty.channel.EventLoopGroup;
+import io.netty.handler.codec.http.HttpHeaders;
 import io.netty.handler.codec.http.HttpMethod;
+import io.netty.handler.proxy.HttpProxyHandler;
 import io.netty.handler.proxy.ProxyConnectException;
 import io.netty.handler.stream.ChunkedNioFile;
 import io.netty.handler.stream.ChunkedStream;
@@ -41,6 +43,8 @@ import reactor.netty.NettyOutbound;
 import reactor.netty.NettyPipeline;
 import reactor.netty.http.client.HttpClientRequest;
 import reactor.netty.http.client.HttpClientResponse;
+import reactor.util.function.Tuple2;
+import reactor.util.function.Tuples;
 
 import javax.net.ssl.SSLException;
 import java.io.IOException;
@@ -50,8 +54,6 @@ import java.nio.file.StandardOpenOption;
 import java.time.Duration;
 import java.util.Objects;
 import java.util.function.BiFunction;
-
-import static com.azure.core.http.netty.implementation.Utility.closeConnection;
 
 /**
  * This class provides a Netty-based implementation for the {@link HttpClient} interface. Creating an instance of this
@@ -121,7 +123,8 @@ class NettyAsyncHttpClient implements HttpClient {
     private Mono<HttpResponse> attemptAsync(HttpRequest request, boolean eagerlyReadResponse,
         boolean ignoreResponseBody, boolean headersEagerlyConverted, Long responseTimeout,
         ProgressReporter progressReporter, boolean proxyRetry) {
-        Flux<HttpResponse> nettyRequest = nettyClient.request(toReactorNettyHttpMethod(request.getHttpMethod()))
+        Flux<Tuple2<HttpResponse, HttpHeaders>> nettyRequest = nettyClient
+            .request(toReactorNettyHttpMethod(request.getHttpMethod()))
             .uri(request.getUrl().toString())
             .send(bodySendDelegate(request))
             .responseConnection(responseDelegate(request, disableBufferCopy, eagerlyReadResponse, ignoreResponseBody,
@@ -133,11 +136,18 @@ class NettyAsyncHttpClient implements HttpClient {
         }
 
         return nettyRequest.single()
-            .flatMap(response -> {
+            .flatMap(responseAndHeaders -> {
+                HttpResponse response = responseAndHeaders.getT1();
                 if (addProxyHandler && response.getStatusCode() == 407) {
-                    return proxyRetry
-                        ? Mono.error(new ProxyConnectException("Connection to proxy failed."))
-                        : Mono.error(new ProxyConnectException("First attempt to connect to proxy failed."));
+                    if (proxyRetry) {
+                        // Exhausted retry attempt return an error.
+                        return Mono.error(new HttpProxyHandler.HttpProxyConnectException(
+                            "Failed to connect to proxy. Status: 407", responseAndHeaders.getT2()));
+                    } else {
+                        // Retry the request.
+                        return attemptAsync(request, eagerlyReadResponse, ignoreResponseBody, headersEagerlyConverted,
+                            responseTimeout, progressReporter, true);
+                    }
                 } else {
                     return Mono.just(response);
                 }
@@ -286,7 +296,7 @@ class NettyAsyncHttpClient implements HttpClient {
      * HttpHeaders.
      * @return a delegate upon invocation setup Rest response object
      */
-    private static BiFunction<HttpClientResponse, Connection, Mono<HttpResponse>> responseDelegate(
+    private static BiFunction<HttpClientResponse, Connection, Mono<Tuple2<HttpResponse, HttpHeaders>>> responseDelegate(
         HttpRequest restRequest, boolean disableBufferCopy, boolean eagerlyReadResponse, boolean ignoreResponseBody,
         boolean headersEagerlyConverted) {
         return (reactorNettyResponse, reactorNettyConnection) -> {
@@ -314,11 +324,11 @@ class NettyAsyncHttpClient implements HttpClient {
                 return Mono.using(() -> reactorNettyConnection, connection -> connection.inbound().receive()
                     .aggregate().asByteArray()
                     .switchIfEmpty(Mono.just(EMPTY_BYTES))
-                    .map(bytes -> new NettyAsyncHttpBufferedResponse(reactorNettyResponse, restRequest, bytes,
-                        headersEagerlyConverted)), Utility::closeConnection);
+                    .map(bytes -> Tuples.of(new NettyAsyncHttpBufferedResponse(reactorNettyResponse, restRequest, bytes,
+                        headersEagerlyConverted), reactorNettyResponse.responseHeaders())), Utility::closeConnection);
             } else {
-                return Mono.just(new NettyAsyncHttpResponse(reactorNettyResponse, reactorNettyConnection, restRequest,
-                    disableBufferCopy, headersEagerlyConverted));
+                return Mono.just(Tuples.of(new NettyAsyncHttpResponse(reactorNettyResponse, reactorNettyConnection,
+                    restRequest, disableBufferCopy, headersEagerlyConverted), reactorNettyResponse.responseHeaders()));
             }
         };
     }