+++ sdk/digitaltwins/azure-digitaltwins-core/CHANGELOG.md
@@ -10,6 +10,16 @@
 
 ### Other Changes
 
+## 1.3.13 (2023-09-22)
+
+### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-core` from `1.42.0` to version `1.43.0`.
+- Upgraded `azure-core-http-netty` from `1.13.6` to version `1.13.7`.
+- Upgraded `azure-core-serializer-json-jackson` from `1.4.3` to version `1.4.4`.
+
 ## 1.3.12 (2023-08-18)
 
 ### Other Changes